requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
biopython==1.83
requests==2.31.0
pytest==7.4.0
# Optional: local LLM via ollama (install only if using ollama backend)
# ollama>=0.1.0
